I created a responsive Quiz web app with Vanilla JS
Hi devs, this is my first project with javascript. I Created a quiz app with allows user to take a test within a time range, allows them to select the option of their choice, click the next button or use the previous button or decide to quit or submit the test. Which then pop up the result summary of the test and an option to retake the test or check answers.